refactor: move v2 to top level
