+++ personal_mnemonic_medium/data_access/ankiconnect_gateway.py
