+++ personal_mnemonic_medium/domain/prompt_destination/anki_connect/anki_prompt_css.py
