+++ personal_mnemonic_medium/domain/prompt_destination/anki_connect/prompt_converter/prompts/anki_qa.py
