+++ personal_mnemonic_medium/domain/prompt_destination/anki_connect/test_ankiconnect_destination.py
