+++ personal_mnemonic_medium/domain/prompt_destination/base_prompt_destination.py
