+++ personal_mnemonic_medium/domain/prompt_source/base_prompt_source.py
