+++ personal_mnemonic_medium/domain/prompt_source/document_ingesters/base_document_ingester.py
