+++ personal_mnemonic_medium/domain/prompt_source/document_ingesters/markdown_document_ingester.py
