+++ personal_mnemonic_medium/domain/prompt_source/document_prompt_source.py
