+++ personal_mnemonic_medium/domain/prompt_source/prompt_extractors/base_prompt_extractor.py
