+++ personal_mnemonic_medium/domain/prompt_source/prompt_extractors/cloze_prompt_extractor.py
