+++ personal_mnemonic_medium/domain/prompt_source/prompt_extractors/qa_prompt_extractor.py
