+++ personal_mnemonic_medium/domain/prompt_source/prompt_extractors/test_cloze_prompt_extractor.py
