+++ personal_mnemonic_medium/domain/prompts/base_prompt.py
