+++ personal_mnemonic_medium/domain/prompts/cloze_prompt.py
