+++ personal_mnemonic_medium/domain/prompts/qa_prompt.py
